+++ cli/src/server.js
@@ -29,6 +29,7 @@ import migrateMongo, { MIGRATIONS_PATH } from 'cli/migrateMongo';
 
 import seed from 'cli/seed';
 import expirationNotificationEmails from 'cli/commands/expirationNotificationEmails';
+import orgUsageTracker from 'cli/commands/orgUsageTracker';
 
 program.version('0.0.1');
 
@@ -57,7 +58,7 @@ program
   .option('-s, --since [since]', 'Requeue statements stored since this ISO date')
   .action(requeueStatements);
 // 1. Stop workers.
-// 2. db.personas.remove({}); db.queryBuilderCaches.remove({}); db.queryBuilderCacheValues.remove({}); db.personaIdentifiers.remove({}); db.statements.update({}, {$set: {processingQueues: [], completedQueues: []}}, {multi: true});
+// 2. db.personas.remove({}); db.queryBuilderCaches.remove({}); db.queryBuilderCacheValues.remove({}); db.personaIdentifiers.remove({}); db.statements.updateMany({}, {$set: {processingQueues: [], completedQueues: []}});
 // 3a. node cli/dist/server updateStatementCount #we skip running the lrs count workers using this instead
 // 3b. node cli/dist/server requeueStatements
 // 4. npm run start-dev-worker
@@ -153,5 +154,9 @@ program
   .action(expirationNotificationEmails)
   .option('--weekBefore [weekBefore]', 'The date of when to send the week before email');
 // node cli/dist/server expirationNotificationEmails
+
+program
+  .command('orgUsageTracker')
+  .action(orgUsageTracker);
 
 program.parse(process.argv);