+++ ui/src/containers/VisualiseResults/CounterResults.js
@@ -12,13 +12,13 @@ const component = args => ({ width, height }) => {
     <Counter
       color={colors.get(0)}
       results={results}
-      maxSize={Math.min(width, height)}
       model={model}
-      width={width} />
+      width={width}
+      height={height} />
   );
 };
 
-const Sizer = (...args) => <AutoSizer>{component(...args)}</AutoSizer>;
+const Sizer = (...args) => <AutoSizer style={{ minHeight: '100%' }}>{component(...args)}</AutoSizer>;
 
 export default compose(
   withStatementsVisualisation,