+++ worker/src/handlers/importPersonas/importPersonaHandler.js
@@ -1,84 +1,23 @@
 import { publish } from 'lib/services/queue';
-import { promisify, map } from 'bluebird';
-import ImportPersonasLock from 'lib/models/importPersonasLock';
-import { getIfis } from 'lib/services/importPersonas/personasImportHelpers';
+import { promisify } from 'bluebird';
 import { PERSONA_IMPORT_QUEUE, STAGE_IMPORTED } from 'lib/constants/personasImport';
 import importPersona from 'lib/services/importPersonas/importPersona';
 import moment from 'moment';
 import PersonasImport from 'lib/models/personasImport';
 import { addErrorsToCsv } from 'lib/services/importPersonas/importPersonas';
 
-const LOCK_TIMEOUT = 120; // seconds
-
-// exported for testing
-export const establishLock = async ({
-  structure,
-  data,
-  organisation,
-  ifis = getIfis({
-    structure,
-    row: data
-  }), // private
-  lockTimeout = LOCK_TIMEOUT * 1000 // testing
-}) => {
-  try {
-    const result = await ImportPersonasLock.create({
-      organisation,
-      ifis
-    });
-
-    return result;
-  } catch (err) {
-    // DuplicateKey
-    if (err.code && err.code === 11000) {
-      const models = ImportPersonasLock.find({
-        organisation,
-        ifis: {
-          $in: ifis
-        }
-      });
-
-      let remainingLocks = false;
-      const deletePromises = map(models, (model) => {
-        if (moment(model.createdAt).add(lockTimeout, 'milliseconds').isBefore(moment())) {
-          return model.remove();
-        }
-        remainingLocks = true;
-      });
-
-      await deletePromises;
-      if (remainingLocks === false) {
-        return await establishLock({
-          structure,
-          data,
-          organisation,
-          ifis,
-          lockTimeout
-        });
-      }
-
-      return false;
-    }
-    throw err;
-  }
-};
-
-const releaseLock = async ({
-  lock
-}) => {
-  await lock.remove();
-};
-
 export const finishedProcessing = async ({
   personaImportId
 }) => {
-  const personasImport = await PersonasImport.findOneAndUpdate({
+  await PersonasImport.updateOne({
     _id: personaImportId
   }, {
     importStage: STAGE_IMPORTED,
     importedAt: moment().toDate()
-  }, {
-    new: true
+  });
+
+  const personasImport = await PersonasImport.findOne({
+    _id: personaImportId
   });
 
   await addErrorsToCsv({
@@ -95,9 +34,18 @@ export default personaService => async ({
   structure,
   organisation
 }, done) => {
-  // establish lock
-  const lock = await establishLock({ structure, data, organisation });
-  if (!lock) {
+  try {
+    const { processedCount, totalCount } = await importPersona({
+      personaImportId,
+      structure,
+      organisation,
+      personaService
+    })(data, index);
+
+    if (totalCount && processedCount >= totalCount) {
+      await finishedProcessing({ personaImportId });
+    }
+  } catch (err) {
     await promisify(publish)({
       queueName: PERSONA_IMPORT_QUEUE,
       payload: {
@@ -108,25 +56,7 @@ export default personaService => async ({
         organisation
       }
     });
-    done();
-    return;
   }
-
-  // Do the magic
-  const { processedCount, totalCount } = await importPersona({
-    personaImportId,
-    structure,
-    organisation,
-    personaService
-  })(data, index);
-
-  // have we finished processing ???
-  if (totalCount && processedCount >= totalCount) {
-    await finishedProcessing({ personaImportId });
-  }
-
-  // release lock
-  await releaseLock({ lock });
 
   done();
 };