+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+# THIS LOG HAS BEEN DEPRECATED AS OF v2.6.4. [PLEASE REFER TO OUR RELEASES INSTEAD.](https://github.com/LearningLocker/learninglocker/releases)
+
 Format based on [Keep a Changelog](http://keepachangelog.com/)
 
 ## [unreleased]
@@ -8,6 +10,23 @@ Format based on [Keep a Changelog](http://keepachangelog.com/)
 ### Fixed
 ### Security
 ### Migrations
+
+## [2.6.4]
+### Added
+  - Adds context activity types to query builder
+### Fixed
+  - Hide server version in headers by default
+
+## [2.6.3]
+### Fixed
+  - Counter scale and centering
+  - Allow merging from non-existent persona
+  - Allow upsert for toPersona on import
+  - Update LRS statement count one by one
+
+## [2.6.2]
+### Fixed
+  - Removed unused nodemon package, solves for flatmap dependency
 
 ## [2.6.1]
 ### Fixed